+++ polaris-core/src/main/java/org/apache/polaris/core/config/PolarisConfiguration.java
@@ -24,8 +24,6 @@ import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * An ABC for Polaris configurations that alter the service's behavior TODO: deprecate unsafe
@@ -34,8 +32,6 @@ import org.slf4j.LoggerFactory;
  * @param <T> The type of the configuration
  */
 public abstract class PolarisConfiguration<T> {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(PolarisConfiguration.class);
 
   private static final List<PolarisConfiguration<?>> allConfigurations = new ArrayList<>();
 
@@ -174,13 +170,16 @@ public abstract class PolarisConfiguration<T> {
     }
 
     /**
-     * Used to support backwards compatibility before there were reserved properties. Usage of this
-     * method should be removed over time.
-     *
-     * @deprecated Use {@link #catalogConfig()} instead.
+     * @deprecated Use {@link #legacyCatalogConfig(String)} instead.
      */
-    @Deprecated
+    @SuppressWarnings("DeprecatedIsStillUsed")
+    @Deprecated(since = "1.0.0-incubating", forRemoval = true)
     public Builder<T> catalogConfigUnsafe(String catalogConfig) {
+      return legacyCatalogConfig(catalogConfig);
+    }
+
+    /** Used to support backwards compatibility before there were reserved properties. */
+    Builder<T> legacyCatalogConfig(String catalogConfig) {
       if (catalogConfig.startsWith(SAFE_CATALOG_CONFIG_PREFIX)) {
         throw new IllegalArgumentException(
             "Unsafe catalog configs are not expected to start with " + SAFE_CATALOG_CONFIG_PREFIX);