+++ runtime/service/src/test/java/org/apache/polaris/service/test/PolarisIntegrationTestFixture.java
@@ -107,7 +107,7 @@ public class PolarisIntegrationTestFixture {
     BasePersistence metaStoreSession =
         helper.metaStoreManagerFactory.getOrCreateSession(realmContext);
     PolarisCallContext polarisContext =
-        new PolarisCallContext(realmContext, metaStoreSession, helper.configurationStore);
+        new PolarisCallContext(realmContext, metaStoreSession, helper.configurationSource);
     PolarisMetaStoreManager metaStoreManager =
         helper.metaStoreManagerFactory.getOrCreateMetaStoreManager(realmContext);
     PrincipalEntity principal = metaStoreManager.findRootPrincipal(polarisContext).orElseThrow();